pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-pipeline"
version = "0.1.0"
description = "Personal Finance Data Pipeline"
authors = [
    {name = "Your Name", email = "youremail@example.com"},
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "selenium>=4.9.0",
    "pandas>=2.0.0",
    "google-cloud-bigquery>=3.11.0",
    "pandas-gbq>=0.19.0",
    "python-dotenv>=1.0.0",
    "webdriver-manager>=4.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.pytest]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
